requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
google-generativeai
requests